% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{list_reference_datasets}
\alias{list_reference_datasets}
\title{List Bundled Reference Datasets}
\usage{
list_reference_datasets()
}
\value{
A data frame with dataset identifiers and versions.
}
\description{
List Bundled Reference Datasets
}
\examples{
list_reference_datasets()
}
