% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{parameter_overrides}
\alias{parameter_overrides}
\title{Parameter Override Audit Trail}
\usage{
parameter_overrides(x)
}
\arguments{
\item{x}{Canonical tables or an \code{rwa_calculation_result}.}
}
\value{
A data frame of old/new values and governance information.
}
\description{
Parameter Override Audit Trail
}
\examples{
parameter_overrides(generate_synthetic_tables(bank_profile = "KSA_BANK"))
}
\seealso{
\code{\link[=override_regulatory_parameters]{override_regulatory_parameters()}}
}
