% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print.rwa_calculation_result}
\alias{print.rwa_calculation_result}
\title{Print a Calculation Result}
\usage{
\method{print}{rwa_calculation_result}(x, ...)
}
\arguments{
\item{x}{An \code{rwa_calculation_result}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Calculation Result
}
