% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{print.rwa_domain_analysis}
\alias{print.rwa_domain_analysis}
\title{Print a Domain Analysis}
\usage{
\method{print}{rwa_domain_analysis}(x, ...)
}
\arguments{
\item{x}{An \code{rwa_domain_analysis}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Domain Analysis
}
