% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print.rwa_validation_report}
\alias{print.rwa_validation_report}
\title{Print a Validation Report}
\usage{
\method{print}{rwa_validation_report}(x, ...)
}
\arguments{
\item{x}{An \code{rwa_validation_report}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Validation Report
}
