% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print.rwa_workspace}
\alias{print.rwa_workspace}
\title{Print a Workspace}
\usage{
\method{print}{rwa_workspace}(x, ...)
}
\arguments{
\item{x}{An \code{rwa_workspace}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Workspace
}
