% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{riskweightedassets-package}
\alias{riskweightedassets}
\alias{riskweightedassets-package}
\title{Reproducible Risk-Weighted Asset Calculations}
\description{
The package implements deterministic, auditable risk-weighted-asset and
capital calculations using canonical data frames or versioned workbooks.
It contains synthetic reference data and source metadata, but no downloaded
regulatory documents and no customer data.
}
\section{Legal information}{

Copyright 2026 RiskDataScience GmbH. GPL-3. Imprint:
\url{https://riskdatascience.net/impressum/}. Privacy policy:
\url{https://riskdatascience.net/datenschutzerklaerung/}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rds0001/risk-weighted-assets-r}
  \item \url{https://riskdatascience.net/impressum/}
  \item \url{https://riskdatascience.net/datenschutzerklaerung/}
  \item Report bugs at \url{https://github.com/rds0001/risk-weighted-assets-r/issues}
}

}
\author{
\strong{Maintainer}: RiskDataScience GmbH \email{riskdatascience@web.de} [copyright holder]

Authors:
\itemize{
  \item RiskDataScience GmbH \email{riskdatascience@web.de} [copyright holder]
}

}
\keyword{internal}
