% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{rwa_validation}
\alias{rwa_validation}
\title{Calculation Validation Report}
\usage{
rwa_validation(result)
}
\arguments{
\item{result}{An \code{rwa_calculation_result}.}
}
\value{
The \code{rwa_validation_report} attached to the result.
}
\description{
Calculation Validation Report
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
r <- calculate_tables(generate_synthetic_tables(bank_profile = "KSA_BANK"))
rwa_validation(r)
\dontshow{\}) # examplesIf}
}
