% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{table_dictionary}
\alias{table_dictionary}
\title{Canonical Table Dictionary}
\usage{
table_dictionary()
}
\value{
A data frame describing all logical tables, workbooks and sheets.
}
\description{
Canonical Table Dictionary
}
\examples{
head(table_dictionary())
}
\seealso{
\code{\link[=table_schema]{table_schema()}}
}
