% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{failed_controls}
\alias{failed_controls}
\title{Failed Calculation Controls}
\usage{
failed_controls(result, view = c("applied", "fully_loaded"))
}
\arguments{
\item{result}{An \code{rwa_calculation_result}.}

\item{view}{\code{"applied"} or \code{"fully_loaded"}.}
}
\value{
A data frame containing only failed controls.
}
\description{
Failed Calculation Controls
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
r <- calculate_tables(generate_synthetic_tables(bank_profile = "KSA_BANK"))
failed_controls(r)
\dontshow{\}) # examplesIf}
}
