% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_api.R
\name{rwa_result_table}
\alias{rwa_result_table}
\title{Read One Result Table}
\usage{
rwa_result_table(result, table, view = c("applied", "fully_loaded"))
}
\arguments{
\item{result}{An \code{rwa_calculation_result}.}

\item{table}{Result-table name.}

\item{view}{\code{"applied"} or \code{"fully_loaded"}.}
}
\value{
A result data frame.
}
\description{
Read One Result Table
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
r <- calculate_tables(generate_synthetic_tables(bank_profile = "KSA_BANK"))
rwa_result_table(r, "Capital_Stack")
\dontshow{\}) # examplesIf}
}
